+++ gameedit.h
@@ -117,7 +117,7 @@ class CGameEditThing : public CThing
 			RFile* pFile,											// In:  File to load from
 			bool bEditMode,										// In:  True for edit mode, false otherwise
 			int16_t sFileCount,										// In:  File count (unique per file, never 0)
-			ULONG	ulFileVersion)									// In:  Version of file format to load.
+			uint32_t	ulFileVersion)									// In:  Version of file format to load.
 			{
 			// Call base class.
 			int16_t	sResult	= CThing::Load(pFile, bEditMode, sFileCount, ulFileVersion);
